requirements.txt
discord.py==0.16.8
gunicorn==19.3.0
requests>=2.7.0
